refactor: 랜딩 페이지 이미지 최적화
- 형식 변경 (png/jpg/webp -> avif)
- 렌더링되는 크기 고려해서 resize
